move LinterState to it's own directory
